move custom notes to separate file
addressing:
C0302:  1,0: Too many lines in module (1013)
