+++ cli/src/katello/client/main.py
@@ -34,6 +34,7 @@ from katello.client.core import (
   package,
   errata,
   system,
+  system_custom_info,
   sync_plan,
   shell_command,
   template,
@@ -196,10 +197,10 @@ def setup_admin(katello_cmd):
         system_cmd.add_command('packages', system.InstalledPackages())
         system_cmd.add_command('add_to_groups', system.AddSystemGroups())
         system_cmd.add_command('remove_from_groups', system.RemoveSystemGroups())
-    system_cmd.add_command('add_custom_info', system.AddCustomInfo())
-    system_cmd.add_command('view_custom_info', system.ViewCustomInfo())
-    system_cmd.add_command('update_custom_info', system.UpdateCustomInfo())
-    system_cmd.add_command('remove_custom_info', system.RemoveCustomInfo())
+    system_cmd.add_command('add_custom_info', system_custom_info.AddCustomInfo())
+    system_cmd.add_command('view_custom_info', system_custom_info.ViewCustomInfo())
+    system_cmd.add_command('update_custom_info', system_custom_info.UpdateCustomInfo())
+    system_cmd.add_command('remove_custom_info', system_custom_info.RemoveCustomInfo())
     katello_cmd.add_command('system', system_cmd)
 
     if mode == 'katello':
